move admin report to admin_report.py
